bingo-backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy

python-jose[cryptography]

python-dotenv

pydantic[email]
email-validator

